Move prune to util; change pre_cfpg return type
the pre_cfpg.from_pg function now returns only the relevant nx graph and
tag set after processing is completed, not all of the intermediates.
